set time to long class

+++ src/assignment/a2/Pizza.java
@@ -2,14 +2,14 @@ package assignment.a2;
 
 public class Pizza {
     int kind;
-    int time;
-    int getTime() {
+    long time;
+    long getTime() {
         return time;
     }
     int getKind() {
         return kind;
     }
-    Pizza(int kind, int time) {
+    Pizza(int kind, long time) {
         this.kind = kind;
         this.time = time;
     }
